+++ src/main/java/org/betonquest/betonquest/compatibility/fabled/FabledClassCondition.java
@@ -1,25 +1,25 @@
-package org.betonquest.betonquest.compatibility.skillapi;
+package org.betonquest.betonquest.compatibility.fabled;
 
-import com.sucy.skill.SkillAPI;
-import com.sucy.skill.api.player.PlayerData;
 import org.betonquest.betonquest.Instruction;
 import org.betonquest.betonquest.api.Condition;
 import org.betonquest.betonquest.api.profiles.Profile;
 import org.betonquest.betonquest.exceptions.InstructionParseException;
+import studio.magemonkey.fabled.Fabled;
+import studio.magemonkey.fabled.api.player.PlayerData;
 
 /**
  * Checks if the player has specific class
  */
 @SuppressWarnings("PMD.CommentRequired")
-public class SkillAPIClassCondition extends Condition {
+public class FabledClassCondition extends Condition {
     private final String className;
 
     private final boolean exact;
 
-    public SkillAPIClassCondition(final Instruction instruction) throws InstructionParseException {
+    public FabledClassCondition(final Instruction instruction) throws InstructionParseException {
         super(instruction, true);
         className = instruction.next();
-        if (!SkillAPI.isClassRegistered(className)) {
+        if (!Fabled.isClassRegistered(className)) {
             throw new InstructionParseException("Class '" + className + "' is not registered");
         }
         exact = instruction.hasArgument("exact");
@@ -27,12 +27,11 @@ public class SkillAPIClassCondition extends Condition {
 
     @Override
     protected Boolean execute(final Profile profile) {
-        final PlayerData data = SkillAPI.getPlayerData(profile.getPlayer());
+        final PlayerData data = Fabled.getData(profile.getPlayer());
         if (exact) {
-            return data.isExactClass(SkillAPI.getClass(className));
+            return data.isExactClass(Fabled.getClass(className));
         } else {
-            return data.isClass(SkillAPI.getClass(className));
+            return data.isClass(Fabled.getClass(className));
         }
     }
-
 }