+++ src/main/java/org/betonquest/betonquest/compatibility/fabled/FabledLevelCondition.java
@@ -1,14 +1,14 @@
-package org.betonquest.betonquest.compatibility.skillapi;
+package org.betonquest.betonquest.compatibility.fabled;
 
-import com.sucy.skill.SkillAPI;
-import com.sucy.skill.api.player.PlayerClass;
-import com.sucy.skill.api.player.PlayerData;
 import org.betonquest.betonquest.Instruction;
 import org.betonquest.betonquest.api.Condition;
 import org.betonquest.betonquest.api.profiles.Profile;
 import org.betonquest.betonquest.exceptions.InstructionParseException;
 import org.betonquest.betonquest.exceptions.QuestRuntimeException;
 import org.betonquest.betonquest.instruction.variable.VariableNumber;
+import studio.magemonkey.fabled.Fabled;
+import studio.magemonkey.fabled.api.player.PlayerClass;
+import studio.magemonkey.fabled.api.player.PlayerData;
 
 import java.util.Optional;
 
@@ -16,15 +16,15 @@ import java.util.Optional;
  * Checks the level of the player in SkillAPI.
  */
 @SuppressWarnings("PMD.CommentRequired")
-public class SkillAPILevelCondition extends Condition {
+public class FabledLevelCondition extends Condition {
     private final String className;
 
     private final VariableNumber level;
 
-    public SkillAPILevelCondition(final Instruction instruction) throws InstructionParseException {
+    public FabledLevelCondition(final Instruction instruction) throws InstructionParseException {
         super(instruction, true);
         className = instruction.next();
-        if (!SkillAPI.isClassRegistered(className)) {
+        if (!Fabled.isClassRegistered(className)) {
             throw new InstructionParseException("Class '" + className + "' is not registered");
         }
         level = instruction.getVarNum();
@@ -32,7 +32,7 @@ public class SkillAPILevelCondition extends Condition {
 
     @Override
     protected Boolean execute(final Profile profile) throws QuestRuntimeException {
-        final PlayerData data = SkillAPI.getPlayerData(profile.getPlayer());
+        final PlayerData data = Fabled.getData(profile.getPlayer());
         final Optional<PlayerClass> playerClass = data
                 .getClasses()
                 .stream()
@@ -40,5 +40,4 @@ public class SkillAPILevelCondition extends Condition {
                 .findAny();
         return playerClass.filter(aClass -> level.getInt(profile) <= aClass.getLevel()).isPresent();
     }
-
 }